Move recovery policy into services
